add option to uninstall libs

+++ src/cluster/docker/remove.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
 
-apt remove --purge docker-ce docker-ce-cli containerd.io -y --allow-change-held-packages
+delete_libs=$1
+
+if [[ "${delete_libs}" == "y" ]]; then
+  apt remove --purge docker-ce docker-ce-cli containerd.io -y --allow-change-held-packages
+fi
