+++ src/cluster/kubernetes/remove.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 
+delete_libs=$1
+
 kubeadm reset -f
-apt remove --purge kubeadm kubelet kubectl -y --allow-change-held-packages
+
+if [[ "${delete_libs}" == "y" ]]; then
+  apt remove --purge kubeadm kubelet kubectl -y --allow-change-held-packages
+fi
